Taskoffer/requirements.txt
fastapi
uvicorn
alembic
sqlalchemy
python-dotenv
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
python-multipart
pwdlib[argon2,bcrypt]
email-validator
phonenumbers

